Added more testing code.
Can chat from Python application through standard input. Currently, the standard input is not async. Thus, it blocks the thread. Since this is only a testing feature, there is no reason to bother with fixing it. However, if we want to read in standard input down the read, we should do it async (asyncio).